+++ py/src/kernels/zip_kernel.rs
@@ -10,8 +10,9 @@ use crate::kernels::{
     molecules::{TanimotoBit, TanimotoReal},
     structures::USAlignKernel,
     protspam::ProtSpamKernel,
+    vectors::{Cosine, L1, L2},
 };
-use crate::utils::{BitFingerprint, RealFingerprint, PdbStructure, SWSequence};
+use crate::utils::{BitFingerprint, RealFingerprint, PdbStructure, SWSequence, Vector};
 use crate::core::_utils::linear_progress_bar;
 
 fn compute<T, K>(data1: Vec<T>, data2: Vec<T>, kernel: K, n_threads: usize, pb: Option<&ProgressBar>) -> Vec<f32>
@@ -118,6 +119,12 @@ pub fn zip_kernel(
             py, PdbStructure, USAlignKernel, args, kwargs; data1, data2; n_threads, pb.as_ref()),
         KernelVariant::ProtSpam => dispatch!(
             py, SWSequence, ProtSpamKernel, args, kwargs; data1, data2; n_threads, pb.as_ref()),
+        KernelVariant::Cosine => dispatch!(
+            py, Vector, Cosine, args, kwargs; data1, data2; n_threads, pb.as_ref()),
+        KernelVariant::L1 => dispatch!(
+            py, Vector, L1, args, kwargs; data1, data2; n_threads, pb.as_ref()),
+        KernelVariant::L2 => dispatch!(
+            py, Vector, L2, args, kwargs; data1, data2; n_threads, pb.as_ref()),
     };
 
     if let Some(ref pb) = pb { pb.finish(); }