+++ refnd/src/kernels/mod.rs
@@ -2,3 +2,4 @@ pub mod molecules;
 pub mod alignments;
 pub mod usalign;
 pub mod protspam;
+pub mod vectors;